chore(chat): remove dead lib/arweave; move uploadFile to lib/files
The api migration broke every direct Arweave consumer in chat. The
remaining callers of lib/arweave/* were three orphans (lib/txtGeneration.ts,
lib/ai/generateImage.ts, lib/email/generateTxtFileEmail.ts) imported by
nothing, plus uploadFile.tsx itself which was repointed to api in this
PR and has nothing Arweave-specific left.

- Move lib/arweave/uploadFile.tsx → lib/files/uploadFile.tsx (and rename
  ArweaveUploadResponse → UploadFileResponse).
- Update the 4 hook imports (useUser, useOrgSettings, useSaveKnowledgeEdit,
  useArtistSetting).
- Delete the orphaned consumers.
- Delete lib/arweave/ entirely.

Net: -415 / +7 lines.
