--- lib/arweave/arweave.ts
@@ -1,5 +1,0 @@
-export type ArweaveURL = `ar://${string}`;
-
-export function isArweaveURL(url: string | null | undefined): boolean {
-  return url && typeof url === "string" ? url.startsWith("ar://") : false;
-}
